+++ test/integration/059_source_overrides_test/test_source_overrides.py
@@ -76,11 +76,6 @@ class TestSourceOverrides(DBTIntegrationTest):
 
     @use_profile('postgres')
     def test_postgres_source_overrides(self):
-        # without running 'deps', our source overrides are invalid
-        _, stdout = self.run_dbt_and_capture(['compile'], strict=False)
-        self.assertRegex(stdout, r'WARNING(\x1b\[0m\])?: During parsing, dbt encountered source overrides that had no target')
-        schema_path = os.path.join('models', 'schema.yml')
-        self.assertIn(f'Source localdep.my_source (in {schema_path})', stdout)
         self.run_dbt(['deps'])
         seed_results = self.run_dbt(['seed'])
         assert len(seed_results) == 5